% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-lm-pruning.R
\name{prune.lm}
\alias{prune.lm}
\title{Prune a lm object.}
\usage{
\method{prune}{lm}(x, ...)
}
\arguments{
\item{x}{A `lm` object returned from [stats::lm()].}
}
\description{
Removes the environment from an lm object.
}
\seealso{
[stats::lm()]
}
\concept{lm pruners}
