% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruner.R
\name{prune_rf}
\alias{prune_rf}
\title{prune big random forest models}
\usage{
prune_rf(object)
}
\arguments{
\item{object}{pre-trained random forest}
}
\value{
size reduced model
}
\description{
Reduce the size of the random forest models trained based on randomForest package
}
